moved graph generation logic into it's own file to make it easier to read
